+++ packages/react/src/badge/docs/code.mdx
@@ -2,6 +2,4 @@
 
 <ComponentPropsTable name="StatusBadge" />
 
-<ComponentPropsTable name="NotificationBadge" />
-
 <ComponentPropsTable name="IndicatorDot" />
