+++ packages/react/src/notification-badge/NotificationBadge.test.tsx
@@ -1,8 +1,12 @@
 import '@testing-library/jest-dom';
 import 'html-validate/jest';
 import { cleanup, render, screen } from '../../../../test-utils';
-import { NotificationBadge, NotificationBadgeProps } from './NotificationBadge';
-import { badgeToneMap, BadgeTone } from './utils';
+import {
+	NotificationBadge,
+	NotificationBadgeProps,
+	badgeToneMap,
+	BadgeTone,
+} from './NotificationBadge';
 
 afterEach(cleanup);
