+++ packages/react/src/notification-badge/docs/code.mdx
@@ -1,0 +1,3 @@
+## Props
+
+<ComponentPropsTable name="NotificationBadge" />
