+++ packages/react/src/notification-badge/index.ts
@@ -1,0 +1,1 @@
+export * from './NotificationBadge';
