+++ packages/react/src/notification-badge/utils.ts
